+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 * File.move construct [CTC-1643]
 * Validate Xill robot when uploading to Xill Server [XSVR-21]
 * Add Date.isBefore and Date.isAfter constructs [CTC-1672]
+* Add Date.fromTimestamp construct [CTC-1616]
 
 ### Change